Implementação do servidor de catálogo terminada

+++ workspace/MobiServer/src/br/unisinos/swe/server/admin/gae/ChannelTransferObjectGAE.java
@@ -26,7 +26,7 @@ public class ChannelTransferObjectGAE {
 	public String mName;
 	
 	@Persistent
-	public Set<Integer> mServicesKeys;
+	public Set<Long> mServicesKeys;
 
 	public ChannelTransferObjectGAE() {
 	}
